requirements.txt
python-telegram-bot==21.6
APScheduler==3.10.4
fastapi==0.115.5
uvicorn==0.32.0
tenacity==9.0.0
pydantic==2.9.2
python-dotenv==1.0.1
aiofiles==24.1.0
